+++ src/bda/plone/orders/transitions.py
@@ -1,5 +1,8 @@
 # -*- coding: utf-8 -*-
+from bda.plone.orders import events
 from bda.plone.orders import interfaces
+from bda.plone.orders.common import BookingData
+from zope.event import notify
 
 
 def transitions_of_main_state(state):
@@ -48,22 +51,49 @@ def transitions_of_salaried_state(state):
     return transitions
 
 
-def do_transition_for(order_state, transition, event=False):
+def do_transition_for(order_state, transition, context=None, request=None):
     """Do any transition for given ``OrderState`` implementation.
 
     This mixes main state and salaried!
     """
     if transition == interfaces.SALARIED_TRANSITION_SALARIED:
         order_state.salaried = interfaces.SALARIED_YES
+
     elif transition == interfaces.SALARIED_TRANSITION_OUTSTANDING:
         order_state.salaried = interfaces.SALARIED_NO
+
     elif transition == interfaces.STATE_TRANSITION_RENEW:
         order_state.state = interfaces.STATE_NEW
+
     elif transition == interfaces.STATE_TRANSITION_PROCESS:
         order_state.state = interfaces.STATE_PROCESSING
+
     elif transition == interfaces.STATE_TRANSITION_FINISH:
         order_state.state = interfaces.STATE_FINISHED
+
     elif transition == interfaces.STATE_TRANSITION_CANCEL:
-        order_state.state = interfaces.STATE_CANCELLED
+        bookings = getattr(order_state, 'bookings', [order_state])
+        for booking_data in bookings:
+
+            if not isinstance(booking_data, BookingData):
+                # It's a booking record from iterating over ``bookings`` in
+                # OrderData. We have to factorize a BookingData object now.
+                booking_data = BookingData(
+                    context=context,
+                    booking=booking_data
+                )
+            booking_attrs = dict(booking_data.booking.attrs.items())
+            # Set state to cancelled. This includes updat of OrderData state
+            # and update of item_stock.
+            booking_data.state = interfaces.STATE_CANCELLED
+            # Send out event, which includes sending a notification mail
+            event = events.BookingCancelledEvent(
+                context=context,
+                request=request,
+                order_uid=booking_attrs['order_uid'],
+                booking_attrs=booking_attrs,
+            )
+            notify(event)
+
     else:
         raise ValueError(u"Invalid transition: %s" % transition)